pre-beta

+++ src/dynarray.c
@@ -1,4 +1,10 @@
-//https://gist.githubusercontent.com/matugm/4708222/raw/ac5a5804c824e18c4ef1b1c760f54686adf602a1/array.c
+/*
+ * dynarray.c
+ * Mon 11 dec 2017 15:31:45
+ * Johan.Nylander@{nbis|nrm}.se
+ * Source: https://gist.githubusercontent.com/matugm/4708222/raw/ac5a5804c824e18c4ef1b1c760f54686adf602a1/array.c
+ *
+ */
 
 #include "dynarray.h"
 #include <stdlib.h>
@@ -21,18 +27,18 @@ void resizeArray(struct ArrayData* array) {
 	array->size *= 2;  // This is the number of elements, don't multiply by sizeof
 }
 
-int addElement(struct ArrayData *array, int number) {
+int addElement(struct ArrayData *array, long int number) {
 	if (array->counter >= array->size) {
 		resizeArray(array);
 	}
 
-	*(array->pointer + array->counter) = number;  // Pointer arithmetic
+	*(array->pointer + array->counter) = number;  // Pointer arithmetic JN: SIGSEGV here on large (array length of 1,033,005,693 characters) data!
 	array->counter += 1;
 
 	return 0;
 }
 
-int getElement(struct ArrayData *array, int index) {
+int getElement(struct ArrayData *array, long int index) {
 	if (array->counter >= array->size) {
 		return -1;
 	}
